Agents/requirements.txt
openai==0.28.0
python-dotenv
langchain
datetime
wikipedia
langchain_experimental
langchain_openai
numexpr